function/requirements.txt
psycopg2-binary==2.8.5
python-http-client==3.2.7
sendgrid==6.3.1
